requirements.txt
asgiref==3.10.0
Django==5.0.14
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.3
requests==2.32.5
psycopg2-binary==2.9.9
django-unfold==0.42.0
gunicorn==21.2.0
whitenoise==6.6.0
